VER 3.6.3
git-svn-id: https://plugins.svn.wordpress.org/itro-popup/trunk@695372 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ uninstall.php
@@ -4,14 +4,21 @@ if ( !defined( 'WP_UNINSTALL_PLUGIN' ) ) {exit ();}
 else
 {
 	include_once ('functions/database-function.php');
-	$preview_id = itro_get_option('preview_id'); //delete preview page
-	wp_delete_post( $preview_id , true );
-	
-	$con = mysql_connect(DB_HOST,DB_USER,DB_PASSWORD);
-	if (!$con){die('Could not connect: ' . mysql_error());}
-	mysql_select_db(DB_NAME);
-	
-	mysql_query('DROP TABLE wp_itro_plugin_option');
-	mysql_query('DROP TABLE wp_itro_plugin_field');
+	if( itro_get_option('delete_data') == 'yes' )
+	{
+		$preview_id = itro_get_option('preview_id'); //delete preview page
+		wp_delete_post( $preview_id , true );
+		
+		delete_option('itro_curr_ver');
+		delete_option('itro_prev_ver');
+		delete_option('delete_data');
+		
+		$con = mysql_connect(DB_HOST,DB_USER,DB_PASSWORD);
+		if (!$con){die('Could not connect: ' . mysql_error());}
+		mysql_select_db(DB_NAME);
+		
+		mysql_query('DROP TABLE wp_itro_plugin_option');
+		mysql_query('DROP TABLE wp_itro_plugin_field');
+	}
 }
 ?>
